able to capture locked color into array
