+++ components/PaletteBox.js
@@ -6,21 +6,21 @@ import ColorBox from './ColorBox';
 
 function PaletteBox() {
   const [colors, setColors] = useState([]);
-  const [isLocked, setIsLocked] = useState(false);
-  const [lockedColors, setLockedColors] = useState([]);
+  // const [isLocked, setIsLocked] = useState(false);
+  // const [lockedColors, setLockedColors] = useState([]);
 
-  const toggleLock = () => {
-    setIsLocked((locked) => {
-      if (!locked) {
-        setLockedColors([colors]);
-        console.warn('registered as:', locked, 'unlocked:', lockedColors);
-      } else {
-        setLockedColors((prevColors) => prevColors.slice(0, -1));
-        console.warn('registered as:', locked, 'locked:', lockedColors);
-      }
-      return !locked;
-    });
-  };
+  // const toggleLock = () => {
+  //   setIsLocked((locked) => {
+  //     if (!locked) {
+  //       setLockedColors([colors]);
+  //       console.warn('registered as:', locked, 'unlocked:', lockedColors);
+  //     } else {
+  //       setLockedColors((prevColors) => prevColors.slice(0, -1));
+  //       console.warn('registered as:', locked, 'locked:', lockedColors);
+  //     }
+  //     return !locked;
+  //   });
+  // };
 
   const randomColor = () => {
     let arr = [];
@@ -33,7 +33,6 @@ function PaletteBox() {
 
   useEffect(() => {
     randomColor();
-    toggleLock();
   }, []);
 
   return (
@@ -42,7 +41,7 @@ function PaletteBox() {
         <div className="palette-box mt-4">
           {
             colors.map((color) => (
-              <ColorBox color={color} isLocked={isLocked} toggleLock={toggleLock} />
+              <ColorBox color={color} />
             ))
           }
         </div>
